refactor: Avoid dependencies to the "o.o.standalone.config" package
DI "global" parameter qualifiers should not be in the config package, but be available in a
"global" package.
