+++ src/main/java/org/opentripplanner/datastore/configure/DataStoreModule.java
@@ -13,7 +13,7 @@ import org.opentripplanner.datastore.api.GoogleStorageDSRepository;
 import org.opentripplanner.datastore.api.OtpDataStoreConfig;
 import org.opentripplanner.datastore.base.DataSourceRepository;
 import org.opentripplanner.datastore.file.FileDataSourceRepository;
-import org.opentripplanner.standalone.config.OtpBaseDirectory;
+import org.opentripplanner.standalone.config.api.OtpBaseDirectory;
 
 /**
  * This is the global access point to create a data store and create datasource objects(tests). It
